Move postmaster into postgresql package
